+++ config/wdio/wdio.conf.js
@@ -3,7 +3,7 @@ const glob = require('glob');
 
 const path = require('path');
 const {
-  Axe: AxeService, SeleniumDocker: SeleniumDockerService, ServeStaticService, Terra: TerraService,
+  SeleniumDocker: SeleniumDockerService, ServeStaticService, Terra: TerraService,
 } = require('../../lib/wdio/services/index');
 const visualRegressionConfig = require('./visualRegressionConf');
 
@@ -23,11 +23,15 @@ const config = {
     path.join('test*', 'wdio', '**', '*-spec.js'),
   ],
   maxInstances: 1,
-  capabilities: [
-    {
-      browserName: 'chrome',
+  capabilities: [{
+    browserName: 'chrome',
+    'goog:chromeOptions': {
+      /** Run in headless mode since Chrome 69 cannot reach the tiny viewport size due to a omnibox size change
+       * made by the chrome team. See https://bugs.chromium.org/p/chromedriver/issues/detail?id=2626#c1.
+       */
+      args: ['headless', 'disable-gpu'],
     },
-  ],
+  }],
 
   sync: true,
   logLevel: 'silent',
@@ -37,7 +41,7 @@ const config = {
   waitforTimeout: 3000,
   connectionRetryTimeout: 90000,
   connectionRetryCount: 1,
-  services: ['visual-regression', AxeService, TerraService, SeleniumDockerService, ServeStaticService],
+  services: ['visual-regression', TerraService, SeleniumDockerService, ServeStaticService],
 
   visualRegression: visualRegressionConfig,
 
@@ -49,7 +53,7 @@ const config = {
   ...locale && { locale },
   formFactor: process.env.FORM_FACTOR || 'huge',
 
-  seleniumVersion: '3.11',
+  seleniumVersion: '3.14',
   seleniumDocker: {
     enabled: !ci,
   },