+++ jestConfig.js
@@ -2,10 +2,12 @@ module.exports = {
   collectCoverage: true,
   collectCoverageFrom: [
     'config/webpack/*.js',
+    'scripts/aggregate-themes/**/*.js',
   ],
   coverageDirectory: 'tests/jest/reports/coverage',
   coverageReporters: [
     'html',
+    'text',
     'lcov',
     'cobertura',
     'text-summary',