+++ tests/jest/webpack.config.test.js
@@ -45,7 +45,7 @@ describe('webpack config', () => {
       expect(config.module.rules).toHaveLength(4);
 
       const cssLoaders = [
-        expect.stringContaining('mini-css-extract-plugin/dist/loader.js'),
+        expect.objectContaining({ loader: expect.stringContaining('mini-css-extract-plugin/dist/loader.js') }),
         expect.objectContaining({ loader: 'css-loader' }),
         expect.objectContaining({ loader: 'postcss-loader' }),
         expect.objectContaining({ loader: 'sass-loader' }),
@@ -160,9 +160,9 @@ describe('webpack config', () => {
       expect(MiniCssExtractPlugin).toBeCalled();
       expect(PostCSSAssetsPlugin).toBeCalled();
 
-      const cleanPluginOptions = expect.objectContaining({ exclude: ['stats.json'] });
+      const cleanPluginOptions = expect.objectContaining({ cleanOnceBeforeBuildPatterns: expect.arrayContaining(['!stats.json']) });
 
-      expect(CleanPlugin).toBeCalledWith(outputPath, cleanPluginOptions);
+      expect(CleanPlugin).toBeCalledWith(cleanPluginOptions);
     });
 
     it('removes devtool option', () => {