+++ tests/wdio/axe-spec.js
@@ -1,6 +1,7 @@
 describe('axe', () => {
   const viewports = Terra.viewports('tiny', 'huge');
   const context = 'html';
+
   it('checks accessibility', () => {
     browser.url('/accessible.html');
     expect(browser.axe({ context })).to.be.accessible();